Example of creating a simple CRUD. https://laravel.com/docs/5.2/quickstart
First create the routing inside "web middleware" that adds CSRF
protection and session handling. The routes handle show all, add,
delete.

The Eloquent model handles delete and creating one new object in the
database automatically with it's own method that points to Query
Builder.

The views points to a master template, and then adds it's own view and
data.